K1: Fix false-positive error due to ProperConstraintProcessing
The fix is mostly disabling ProperTypeInferenceConstraintsProcessing
in K1.
It's necessary because that "feature" is incompatible with the state of
other things in K1.
(like not having fork-points in inference there, see the comment at kt52782.kt),
At the same time, it works just fine for K2 and it's forcefully
already enabled there.

^KT-52782 Fixed